# packages/api-client/openapi-spec.yaml
openapi: 3.0.0
paths:
...
  /api/v1/mentor/profile:
    post:
      operationId: createMentorProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMentorProfileDto'
      responses:
        '201':
          description: 멘토 프로필 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorProfileDto'
        '400':
          description: 잘못된 요청 데이터
        '404':
          description: 사용자 또는 태그를 찾을 수 없음
        '409':
          description: 이미 멘토 프로필이 존재함
      summary: 새 멘토 프로필 생성
      tags: &ref_8
        - Mentor
...
  /api/v1/onboarding/profile:
    post:
      operationId: createBasicProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBasicProfileDto'
      responses:
        '200':
          description: 기본 프로필이 성공적으로 생성되었습니다.
          content:
            application/json:
              schema:
                title: UserProfileResponse
                type: object
                properties:
                  message:
                    type: string
                    description: 응답 메시지
                    example: 성공적으로 처리되었습니다.
                  data:
                    $ref: '#/components/schemas/UserProfileDto'
        '400':
          description: 잘못된 요청 데이터
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: 잘못된 요청 데이터입니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '401':
          description: 인증이 필요합니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: 인증이 필요합니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '403':
          description: 접근 권한이 없습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 403
                  message:
                    type: string
                    example: 해당 작업을 수행할 권한이 없습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '409':
          description: 닉네임이 중복되거나 이미 프로필이 존재합니다.
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  message:
                    type: string
                    example: 서버 내부 오류가 발생했습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
      summary: 온보딩 - 기본 프로필 생성
      tags: &ref_9
        - Onboarding
  /api/v1/onboarding/baekjoon/verify/start:
    post:
      operationId: startBaekjoonVerification
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartBaekjoonVerificationDto'
      responses:
        '200':
          description: 백준 인증 세션이 성공적으로 시작되었습니다.
          content:
            application/json:
              schema:
                title: VerificationStatusResponseWrapper
                type: object
                properties:
                  message:
                    type: string
                    description: 응답 메시지
                    example: 성공적으로 처리되었습니다.
                  data:
                    $ref: '#/components/schemas/VerificationStatusResponseDto'
        '400':
          description: 잘못된 요청 데이터
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: 잘못된 요청 데이터입니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '401':
          description: 인증이 필요합니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: 인증이 필요합니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '403':
          description: 접근 권한이 없습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 403
                  message:
                    type: string
                    example: 해당 작업을 수행할 권한이 없습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '404':
          description: 백준 사용자를 찾을 수 없습니다.
        '408':
          description: 외부 API 응답 시간이 초과되었습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 408
                  message:
                    type: string
                    example: 외부 API 응답 시간이 초과되었습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '409':
          description: 이미 진행 중인 인증이 있거나, 다른 사용자가 사용중인 핸들입니다.
        '429':
          description: API 호출 한도를 초과했습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 429
                  message:
                    type: string
                    example: API 호출 한도를 초과했습니다. 잠시 후 다시 시도해주세요.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  message:
                    type: string
                    example: 서버 내부 오류가 발생했습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '503':
          description: 외부 서비스를 사용할 수 없습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  message:
                    type: string
                    example: 외부 서비스를 일시적으로 사용할 수 없습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
      summary: 온보딩 - 백준 인증 시작
      tags: *ref_9
  /api/v1/onboarding/baekjoon/verify/complete:
    post:
      operationId: completeBaekjoonVerification
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteBaekjoonVerificationDto'
      responses:
        '200':
          description: 백준 인증이 성공적으로 완료되었습니다.
          content:
            application/json:
              schema:
                title: BaekjoonVerificationResultResponseWrapper
                type: object
                properties:
                  message:
                    type: string
                    description: 응답 메시지
                    example: 성공적으로 처리되었습니다.
                  data:
                    $ref: '#/components/schemas/BaekjoonVerificationResultResponseDto'
        '400':
          description: 잘못된 요청 데이터
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: 잘못된 요청 데이터입니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '401':
          description: 인증이 필요합니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: 인증이 필요합니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '403':
          description: 접근 권한이 없습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 403
                  message:
                    type: string
                    example: 해당 작업을 수행할 권한이 없습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '404':
          description: 진행 중인 인증 세션을 찾을 수 없습니다.
        '408':
          description: 외부 API 응답 시간이 초과되었습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 408
                  message:
                    type: string
                    example: 외부 API 응답 시간이 초과되었습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '429':
          description: API 호출 한도를 초과했습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 429
                  message:
                    type: string
                    example: API 호출 한도를 초과했습니다. 잠시 후 다시 시도해주세요.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  message:
                    type: string
                    example: 서버 내부 오류가 발생했습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '503':
          description: 외부 서비스를 사용할 수 없습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  message:
                    type: string
                    example: 외부 서비스를 일시적으로 사용할 수 없습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
      summary: 온보딩 - 백준 인증 완료
      tags: *ref_9
  /api/v1/onboarding/analyze:
    post:
      operationId: analyzeSkills
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeSkillsDto'
      responses:
        '200':
          description: 실력 분석이 성공적으로 완료되었습니다.
          content:
            application/json:
              schema:
                title: TagStatisticsOutputResponse
                type: object
                properties:
                  message:
                    type: string
                    description: 응답 메시지
                    example: 성공적으로 처리되었습니다.
                  data:
                    $ref: '#/components/schemas/TagStatisticsOutputDto'
        '400':
          description: 잘못된 요청 데이터
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: 잘못된 요청 데이터입니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '401':
          description: 인증이 필요합니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: 인증이 필요합니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '403':
          description: 접근 권한이 없습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 403
                  message:
                    type: string
                    example: 해당 작업을 수행할 권한이 없습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '404':
          description: 백준 프로필을 찾을 수 없습니다.
        '408':
          description: 외부 API 응답 시간이 초과되었습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 408
                  message:
                    type: string
                    example: 외부 API 응답 시간이 초과되었습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '429':
          description: API 호출 한도를 초과했습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 429
                  message:
                    type: string
                    example: API 호출 한도를 초과했습니다. 잠시 후 다시 시도해주세요.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  message:
                    type: string
                    example: 서버 내부 오류가 발생했습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '503':
          description: 외부 서비스를 사용할 수 없습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  message:
                    type: string
                    example: 외부 서비스를 일시적으로 사용할 수 없습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
      summary: 온보딩 - 실력 분석 실행
      tags: *ref_9
  /api/v1/onboarding/mentor/eligibility:
    get:
      operationId: checkMentorEligibility
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckMentorEligibilityDto'
      responses:
        '200':
          description: 멘토 자격 요건을 성공적으로 확인했습니다.
          content:
            application/json:
              schema:
                title: MentorEligibilityResponse
                type: object
                properties:
                  message:
                    type: string
                    description: 응답 메시지
                    example: 성공적으로 처리되었습니다.
                  data:
                    $ref: '#/components/schemas/MentorEligibilityDto'
        '400':
          description: 잘못된 요청 데이터
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: 잘못된 요청 데이터입니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '401':
          description: 인증이 필요합니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: 인증이 필요합니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '403':
          description: 접근 권한이 없습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 403
                  message:
                    type: string
                    example: 해당 작업을 수행할 권한이 없습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '404':
          description: 백준 프로필을 찾을 수 없습니다.
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  message:
                    type: string
                    example: 서버 내부 오류가 발생했습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
      summary: 온보딩 - 멘토 자격 요건 확인
      tags: *ref_9
  /api/v1/onboarding/mentor/tags:
    get:
      operationId: getMentorTags
      parameters: []
      responses:
        '200':
          description: 멘토 태그 목록을 성공적으로 조회했습니다.
          content:
            application/json:
              schema:
                title: MentorTagsResponseWrapper
                type: object
                properties:
                  message:
                    type: string
                    description: 응답 메시지
                    example: 성공적으로 처리되었습니다.
                  data:
                    $ref: '#/components/schemas/MentorTagsResponseDto'
        '400':
          description: 잘못된 요청 데이터
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: 잘못된 요청 데이터입니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '401':
          description: 인증이 필요합니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: 인증이 필요합니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '403':
          description: 접근 권한이 없습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 403
                  message:
                    type: string
                    example: 해당 작업을 수행할 권한이 없습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  message:
                    type: string
                    example: 서버 내부 오류가 발생했습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
      summary: 온보딩 - 멘토 프로필 태그 목록 조회
      tags: *ref_9
  /api/v1/onboarding/mentor-profile:
    post:
      operationId: createMentorProfileOnboarding
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingCreateMentorProfileDto'
      responses:
        '200':
          description: 멘토 프로필이 성공적으로 생성/업데이트되었습니다.
          content:
            application/json:
              schema:
                title: MentorProfileResponse
                type: object
                properties:
                  message:
                    type: string
                    description: 응답 메시지
                    example: 성공적으로 처리되었습니다.
                  data:
                    $ref: '#/components/schemas/MentorProfileDto'
        '400':
          description: 잘못된 요청 데이터
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: 잘못된 요청 데이터입니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '401':
          description: 인증이 필요합니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: 인증이 필요합니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '403':
          description: 접근 권한이 없습니다
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 403
                  message:
                    type: string
                    example: 해당 작업을 수행할 권한이 없습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
        '404':
          description: 사용자를 찾을 수 없습니다.
        '409':
          description: 멘토 프로필이 이미 존재합니다.
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  message:
                    type: string
                    example: 서버 내부 오류가 발생했습니다.
                  timestamp:
                    type: string
                    example: '2024-01-01T00:00:00.000Z'
      summary: 온보딩 - 멘토 프로필 생성/업데이트
      tags: *ref_9
...
info:
  title: CotePT API
  description: 코딩 테스트 멘토링 플랫폼 API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: JWT 토큰을 입력하세요
      in: header
  schemas:
...
    CreateMentorProfileDto:
      type: object
      properties:
        userId:
          type: string
          description: 연결된 유저의 고유 ID
          example: cld2clvaw00003575n5da3kch
        introductionTitle:
          type: string
          description: 멘토링 소개 제목
          example: 실전처럼 배우는 프론트엔드 테스트 코드 작성법
        introductionContent:
          type: string
          description: 멘토링 소개 내용 (마크다운)
        tagIds:
          description: 멘토의 기술 스택 및 직무 태그 ID 목록
          example:
            - 1
            - 5
            - 10
          type: array
          items:
            type: number
      required:
        - userId
        - tagIds
...
    MentorProfileDto:
      type: object
      properties:
        idx:
          type: number
          description: 멘토 프로필 IDX
          example: 1
        userId:
          type: string
          description: 연결된 유저의 고유 ID
          example: cld2clvaw00003575n5da3kch
        introductionTitle:
          type: string
          description: 멘토링 소개 제목
          example: 실전처럼 배우는 프론트엔드 테스트 코드 작성법
        introductionContent:
          type: string
          description: 멘토링 소개 내용 (마크다운)
        baekjoonTierDisplay:
          type: boolean
          description: 백준 티어 표시 여부
          example: true
        mentoringCount:
          type: number
          description: 총 멘토링 횟수
          example: 15
        totalReviewCount:
          type: number
          description: 총 리뷰 개수
          example: 12
        averageRating:
          type: number
          description: 평균 평점 (0.0 ~ 5.0)
          example: 4.8
        isVerified:
          type: boolean
          description: 멘토 인증 여부
          example: true
        isActive:
          type: boolean
          description: 멘토 활동 여부
          example: true
        profileCompletion:
          type: number
          description: 프로필 완성도 (0 ~ 100)
          example: 95
        tags:
          description: 멘토 태그 목록
          type: array
          items:
            $ref: '#/components/schemas/MentorTagDto'
        createdAt:
          format: date-time
          type: string
          description: 생성일
        updatedAt:
          format: date-time
          type: string
          description: 수정일
      required:
        - idx
        - userId
        - baekjoonTierDisplay
        - mentoringCount
        - totalReviewCount
        - averageRating
        - isVerified
        - isActive
        - profileCompletion
        - tags
        - createdAt
        - updatedAt
...
    CreateBasicProfileDto:
      type: object
      properties:
        userId:
          type: string
          description: 사용자 로그인 아이디 (6~20자, 영문/숫자/특수문자)
          example: dudtod1596
        nickname:
          type: string
          description: 닉네임 (2~50자, 한글/영문만 허용)
          example: 코딩마스터
        profileImageUrl:
          type: string
          description: 프로필 이미지 URL
          example: https://example.com/profile/image.jpg
      required:
        - userId
        - nickname
...
    StartBaekjoonVerificationDto:
      type: object
      properties:
        userId:
          type: string
          description: 사용자 ID
          example: user123
        baekjoonHandle:
          type: string
          description: 백준 핸들
          example: solved_user123
      required:
        - userId
        - baekjoonHandle
...
    CompleteBaekjoonVerificationDto:
      type: object
      properties:
        userId:
          type: string
          description: 사용자 ID
          example: user123
        baekjoonHandle:
          type: string
          description: 백준 핸들
          example: solved_user123
        verificationSessionId:
          type: string
          description: 인증 세션 ID
          example: 550e8400-e29b-41d4-a716-446655440000
      required:
        - userId
        - baekjoonHandle
        - verificationSessionId
...